Add script to clean uri handlers
Mac has a habit of registering any .app opened once as URI handler, and
sometimes picking it as the default handler for the type. This messes
with default browser selection when you e.g. build a bundle for
development purposes in your checkout and run it once just for testing.

This script can be called after development to clean the URI handlers
list a bit as needed.